requirements.txt
## 必要パッケージ(requirements.txt)

streamlit==1.32.0
langchain==0.1.13
langchain-openai==0.1.0
openai>=1.10.0
python-dotenv==1.0.1
pandas==2.2.0
plotly==5.19.0
